+++ tests/unit/Grav/Common/UtilsTest.php
@@ -152,13 +152,36 @@ class UtilsTest extends \Codeception\TestCase\Test
 
     }
 
-    public function testGetMimeType()
+    public function testGetMimeByExtension()
     {
-        $this->assertEquals('application/octet-stream', Utils::getMimeType(''));
-        $this->assertEquals('image/jpeg', Utils::getMimeType('jpg'));
-        $this->assertEquals('image/png', Utils::getMimeType('png'));
-        $this->assertEquals('text/plain', Utils::getMimeType('txt'));
-        $this->assertEquals('application/msword', Utils::getMimeType('doc'));
+        $this->assertEquals('application/octet-stream', Utils::getMimeByExtension(''));
+        $this->assertEquals('text/html', Utils::getMimeByExtension('html'));
+        $this->assertEquals('application/json', Utils::getMimeByExtension('json'));
+        $this->assertEquals('application/atom+xml', Utils::getMimeByExtension('atom'));
+        $this->assertEquals('application/rss+xml', Utils::getMimeByExtension('rss'));
+        $this->assertEquals('image/jpeg', Utils::getMimeByExtension('jpg'));
+        $this->assertEquals('image/png', Utils::getMimeByExtension('png'));
+        $this->assertEquals('text/plain', Utils::getMimeByExtension('txt'));
+        $this->assertEquals('application/msword', Utils::getMimeByExtension('doc'));
+        $this->assertEquals('application/octet-stream', Utils::getMimeByExtension('foo'));
+        $this->assertEquals('foo/bar', Utils::getMimeByExtension('foo', 'foo/bar'));
+        $this->assertEquals('text/html', Utils::getMimeByExtension('foo', 'text/html'));
+    }
+
+    public function testGetExtensionByMime()
+    {
+        $this->assertEquals('html', Utils::getExtensionByMime('*/*'));
+        $this->assertEquals('html', Utils::getExtensionByMime('text/*'));
+        $this->assertEquals('html', Utils::getExtensionByMime('text/html'));
+        $this->assertEquals('json', Utils::getExtensionByMime('application/json'));
+        $this->assertEquals('atom', Utils::getExtensionByMime('application/atom+xml'));
+        $this->assertEquals('rss', Utils::getExtensionByMime('application/rss+xml'));
+        $this->assertEquals('jpg', Utils::getExtensionByMime('image/jpeg'));
+        $this->assertEquals('png', Utils::getExtensionByMime('image/png'));
+        $this->assertEquals('txt', Utils::getExtensionByMime('text/plain'));
+        $this->assertEquals('doc', Utils::getExtensionByMime('application/msword'));
+        $this->assertEquals('html', Utils::getExtensionByMime('foo/bar'));
+        $this->assertEquals('baz', Utils::getExtensionByMime('foo/bar', 'baz'));
     }
 
     public function testNormalizePath()